backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
supabase
opencv-python
face_recognition
numpy
pillow